+++ tests/api/test_statements.py
@@ -4,6 +4,7 @@ from importlib import reload
 
 from ralph import conf
 from ralph.api.routers import statements
+from ralph.backends.database.clickhouse import ClickHouseDatabase
 from ralph.backends.database.es import ESDatabase
 from ralph.backends.database.mongo import MongoDatabase
 
@@ -24,3 +25,8 @@ def test_api_statements_backend_instance_with_runserver_backend_env(monkeypatch)
     monkeypatch.setenv("RALPH_RUNSERVER_BACKEND", "es")
     reload(conf)
     assert isinstance(reload(statements).DATABASE_CLIENT, ESDatabase)
+
+    # ClickHouse backend
+    monkeypatch.setenv("RALPH_RUNSERVER_BACKEND", "clickhouse")
+    reload(conf)
+    assert isinstance(reload(statements).DATABASE_CLIENT, ClickHouseDatabase)