+++ tests/conftest.py
@@ -6,6 +6,7 @@ from .fixtures import hypothesis_configuration  # noqa: F401
 from .fixtures import hypothesis_strategies  # noqa: F401
 from .fixtures.auth import auth_credentials  # noqa: F401
 from .fixtures.backends import (  # noqa: F401
+    clickhouse,
     es,
     es_data_stream,
     es_forwarding,